refactor: change design of donor form page for consistency, add image functionality pending
