+++ lib/screens/donor_new/donor_donations.dart
@@ -9,45 +9,35 @@ import 'package:intl/intl.dart';
 import 'package:donation_app/providers/donation_provider.dart';
 import 'package:donation_app/models/donation_model.dart';
 
-class OrgHomePage extends StatefulWidget {
-  const OrgHomePage({super.key});
+class DonorDonationsPage extends StatefulWidget {
+  const DonorDonationsPage({super.key});
 
   @override
-  _OrgHomePageState createState() => _OrgHomePageState();
+  _DonorDonationsPageState createState() => _DonorDonationsPageState();
 }
 
-class _OrgHomePageState extends State<OrgHomePage> {
+class _DonorDonationsPageState extends State<DonorDonationsPage> {
   late User? _currentUser;
-  Map<String, dynamic>? _userDetails;
 
   @override
   void initState() {
     super.initState();
-    // TODO: Remove, move implementation to org profile page
     // fetch user details
     _currentUser = FirebaseAuth.instance.currentUser;
     if (_currentUser != null) {
-      _fetchUserDetails(); // TODO: Remove, move implementation to org profile page
       // execute initialization of the stream after the layout is completed
       WidgetsBinding.instance.addPostFrameCallback((_) {
-        context.read<DonationProvider>().fetchDonationsToOrg(_currentUser!.uid);
+        context
+            .read<DonationProvider>()
+            .fetchDonationsByDonor(_currentUser!.uid);
       });
     }
   }
 
-  // TODO: Remove, move implementation to org profile page
-  Future<void> _fetchUserDetails() async {
-    final details =
-        await context.read<MyAuthProvider>().getUserDetails(_currentUser!.uid);
-    setState(() {
-      _userDetails = details;
-    });
-  }
-
-  Future<String> _fetchDonorName(String donorId) async {
+  Future<String> _fetchOrgName(String orgId) async {
     final _userDetails =
-        await context.read<MyAuthProvider>().getUserDetails(donorId);
-    return _userDetails['name'] as String? ?? 'Unknown Donor';
+        await context.read<MyAuthProvider>().getUserDetails(orgId);
+    return _userDetails['name'] as String? ?? 'Unknown Organization';
   }
 
   Widget _buildDonationList(List<Donation> donations, String status) {
@@ -69,22 +59,22 @@ class _OrgHomePageState extends State<OrgHomePage> {
       itemBuilder: (context, index) {
         Donation donation = filteredDonations[index];
         return FutureBuilder<String>(
-          future: _fetchDonorName(donation.donorId),
+          future: _fetchOrgName(donation.orgId),
           builder: (context, snapshot) {
             if (snapshot.connectionState == ConnectionState.waiting) {
               return const LinearProgressIndicator();
             } else if (snapshot.hasError) {
               return Text('Error encountered! ${snapshot.error}');
             }
-            String donorName = snapshot.data ?? 'Donor';
-            return _buildDonationCard(donation, donorName);
+            String orgName = snapshot.data ?? 'Organization';
+            return _buildDonationCard(donation, orgName);
           },
         );
       },
     );
   }
 
-  Widget _buildDonationCard(Donation donation, String donorName) {
+  Widget _buildDonationCard(Donation donation, String orgName) {
     return Padding(
       padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 16),
       child: InkWell(
@@ -105,7 +95,7 @@ class _OrgHomePageState extends State<OrgHomePage> {
           surfaceTintColor: Colors.transparent,
           child: ListTile(
             title: Text(
-              donorName, // display the fetched donor name
+              orgName, // display the fetched donor name
               style: const TextStyle(
                 fontSize: 24,
                 fontWeight: FontWeight.bold,
@@ -164,7 +154,7 @@ class _OrgHomePageState extends State<OrgHomePage> {
   Widget build(BuildContext context) {
     // access donations in the provider
     Stream<QuerySnapshot> donationsStream =
-        context.watch<DonationProvider>().donationsToOrg;
+        context.watch<DonationProvider>().donationsByDonor;
 
     final ColorScheme colorScheme = Theme.of(context).colorScheme;
